Reorganize HasFieldMut into its own module
