+++ crates/cgp-field/src/lib.rs
@@ -5,5 +5,5 @@ pub mod traits;
 pub mod types;
 
 pub use cgp_field_macro::{symbol, HasField};
-pub use traits::has_field::{HasField, HasFieldMut};
+pub use traits::{HasField, HasFieldMut};
 pub use types::Char;
