+++ crates/cgp-field/src/traits/mod.rs
@@ -1,1 +1,5 @@
 pub mod has_field;
+pub mod has_field_mut;
+
+pub use has_field::HasField;
+pub use has_field_mut::HasFieldMut;
